@jakobdreyer 2018-09-07
Hi Nikolay

I am working on making it more compatible with your previous version. So I will make another PR soon. I am just working on getting the test_entropy.py ready.

BTW: I seem to get some errors on code-coverage so I wasn’t sure if the PR actually came through.

Cheers
Jakob.

From: Nikolay Donets [mailto:notifications@github.com]
Sent: 7. september 2018 15:35
To: nikdon/pyEntropy <pyEntropy@noreply.github.com>
Cc: Jakob Kisbye Dreyer <JAKD@lundbeck.com>; Author <author@noreply.github.com>
Subject: Re: [nikdon/pyEntropy] Update sample entropy (#8)


@nikdon approved this pull request.

Hi Jacob,

Thank you for the PR, it's very useful. I'll go through it and after that merge

________________________________

In pyentrp/entropy.py<https://github.com/nikdon/pyEntropy/pull/8#discussion_r215953804>:

> @@ -102,60 +101,88 @@ def shannon_entropy(time_series):

     return ent





-def sample_entropy(time_series, sample_length, tolerance=None):

-    """Calculate and return Sample Entropy of the given time series.

-    Distance between two vectors defined as Euclidean distance and can

-    be changed in future releases

-

+def sample_entropy(time_series, m, tol = None):

+    """Calculates the sample entropy of degree m of a time_series.

+

+    This method uses chebychev norm and searches iteratively. It is quite fast for random data, but can be slower is there is structure in the time series.

Possibly a typo here: slower if there is a structure?

________________________________

In pyentrp/entropy.py<https://github.com/nikdon/pyEntropy/pull/8#discussion_r215954140>:

>      Args:

-        time_series: Vector or string of the sample data

-        sample_length: Number of sequential points of the time series

-        tolerance: Tolerance (default = 0.1...0.2 * std(time_series))

-

-    Returns:

-        Vector containing Sample Entropy (float)

-

+        time_series: numpy array of time series

+        m: length of template vector

+        r: tolerance (defaults to 0.1 * std(time_series)))

A typo? Should be tol?

________________________________

In pyentrp/entropy.py<https://github.com/nikdon/pyEntropy/pull/8#discussion_r215958935>:

> +            "Hitlist is bool list and true where next time_series elements match next template elements"

+            hitlist = np.abs(nextcandidates - template[length]) < tol

+            "reduce the search list to those elements that surviced this round"

+            searchlist = nextindxlist[hitlist]

+

+            if length == m-1:

+                Ntemp += np.sum(hitlist)

+            elif length == m :

+                Ntemp_plus += np.sum(hitlist)

+

+

+            length += 1;

+            go = any(hitlist) and length < m + 1

+

+

+    #print('Ntemp m', Ntemp)

Let's remove it? Seem like it's not necessary to keep comments

________________________________

In tests/test_entropy.py<https://github.com/nikdon/pyEntropy/pull/8#discussion_r215959390>:

>

-    def test_multiScaleEntropy(self):

-        ts = TS_SAMPLE_ENTROPY

-        std_ts = np.std(ts)

-        multi_scale_entropy = ent.multiscale_entropy(ts, 4, 0.2 * std_ts)

-        np.testing.assert_array_equal(np.round(multi_scale_entropy, 4), np.array([2.2119, 2.6221, 1.7473, 1.9902]))

+    def test_multiScaleEntropy(self):

+        multi_scale_entropy = ent.multiscale_entropy(RANDOM_TIME_SERIES, 3, maxscale = 4 )

Formatting: redundant space

—
You are receiving this because you authored the thread.
Reply to this email directly, view it on GitHub<https://github.com/nikdon/pyEntropy/pull/8#pullrequestreview-153335488>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AilFeFZQIWn7R6WqAKV0I-rVX-TN2ykCks5uYnYTgaJpZM4Wekc1>.