@jakobdreyer 2018-09-10
Hi Nikolay

The latest commits have changed the input and output interfaces to be same as in your first version. So sample_entropy(dat, sample_length, 0.1) returns a vector of length sample_length of which the first element is = #templates_of_length_1 / [n*(n-1)/2]. 

I have also changed the test_script so that it is reflecting these changes. 

J.