@jakobdreyer 2018-09-10
Hi Nikolay

Let me know if you would like a new 'clean' PR that contains the latest updates (which are contained in the latest commits to this branch).