@nikdon 2018-09-11
Hi Jacob,

Thank you for the contribution. Actually, all commits can be squashed into the one, so it is not necessary to open a new PR.

I will prepare a new release shortly with your changes.